+++ src/main/java/com/f_log/flog/diet/dto/UpdateDietRequest.java
@@ -9,13 +9,13 @@ import java.time.LocalDate;
 @Getter
 @Builder
 public class UpdateDietRequest {
-    private int totalCarbohydrate;
-    private int totalProtein;
-    private int totalFat;
-    private int totalSodium;
-    private int totalCholesterol;
-    private int totalSugars;
-    private int totalCalories;
+    private double totalCarbohydrate;
+    private double totalProtein;
+    private double totalFat;
+    private double totalSodium;
+    private double totalCholesterol;
+    private double totalSugars;
+    private double totalCalories;
     private MealType mealType;
     private LocalDate mealDate;
 }
